+++ fj-doc-freemarker/src/main/java/org/fugerit/java/doc/freemarker/tool/verify/VerifyTemplateInfo.java
@@ -1,0 +1,21 @@
+package org.fugerit.java.doc.freemarker.tool.verify;
+
+import lombok.*;
+
+@ToString
+@AllArgsConstructor
+@RequiredArgsConstructor
+public class VerifyTemplateInfo {
+
+    @Getter
+    @NonNull
+    private Integer resultCode;
+
+    @Getter
+    @NonNull
+    private String templateId;
+
+    @Getter
+    private Exception exception;
+
+}